+++ src/interfaces/Card/NormalReviewCardInterface.ts
@@ -23,6 +23,7 @@ export type MoveType = 'SMALL' | 'HOME' | 'OFFICE' | 'APPOINTMENT' | 'WAITING';
 
 export interface ReviewCardProps {
   type: 'ABLE' | 'MY';
+  myReview: MyReviewItem;
 }
 
 export interface ReviewCardReviews {
@@ -32,28 +33,26 @@ export interface ReviewCardReviews {
   createdAt: string;
 }
 
+export interface MyReviewItem {
+  id: string;
+  createdAt: string;
+  comment: string;
+  score: number;
+  driver: {
+    name: string;
+    image: string;
+  };
+  estimation: {
+    price: number;
+    moveInfos: {
+      serviceType: 'SMALL' | 'HOME' | 'OFFICE';
+      date: string;
+      isSpecificRequest: boolean;
+    };
+  };
+}
+
 export interface MyReviews {
   totalCount: number;
-  list: [
-    {
-      id: string;
-      createdAt: string;
-      comment: string;
-      score: number;
-
-      driver: {
-        name: string;
-        image: string;
-      };
-
-      estimation: {
-        price: number;
-        moveInfos: {
-          serviceType: 'SMALL' | 'HOME' | 'OFFICE';
-          date: string;
-          isSpecificRequest: boolean;
-        };
-      };
-    },
-  ];
+  list: MyReviewItem[];
 }
